+++ test/unit/modules/JSX/index.test.ts
@@ -1,10 +1,17 @@
 import {describe, it, expect} from 'vitest';
 import {nonce} from '../../../../lib/modules/JSX/nonce/use';
+import {Script} from '../../../../lib/modules/JSX/script/Script';
 import {Style} from '../../../../lib/modules/JSX/style/Style';
 import {createCss} from '../../../../lib/modules/JSX/style/use';
 import * as Module from '../../../../lib/modules/JSX/index';
 
 describe('Modules - JSX', () => {
+    describe('Script', () => {
+        it('Should link to the correct module', () => {
+            expect(Module.Script).toEqual(Script);
+        });
+    });
+
     describe('Style', () => {
         it('Should link to the correct module', () => {
             expect(Module.Style).toEqual(Style);